// src/generated/responses.ts
export type GetBrowseCategoriesIdData = GetBrowseCategoriesidData;
export type GetBrowseRecommendationsData = string;
// ...
export interface Icon {
  url: string;
  height: number | null;
  width: number | null;
}
// ...
export interface ItemElement {
  album_type: AlbumAlbumType;
  artists: Owner[];
  external_urls: ExternalUrls;
  href: string;
  id: string;
  images: Icon[];
  is_playable?: boolean;
  name: string;
  release_date: string;
  release_date_precision: ReleaseDatePrecision;
  total_tracks: number | null;
  type: AlbumAlbumType;
  uri: string;
  available_markets?: Market[];
  restrictions?: PlayabilityClass;
}
// ...
export interface GetBrowseCategoriesData {
  categories: Categories;
}

export interface Categories {
  href: string;
  items: GetBrowseCategoriesidData[];
  limit: number;
  next: string;
  offset: number;
  previous: null;
  total: number;
}

export interface GetBrowseCategoriesidData {
  href: string;
  id: string;
  icons: Icon[];
  name: string;
}

export interface GetBrowseFeaturedPlaylistsData {
  message: string;
  playlists: GetBrowseFeaturedPlaylistsDataPlaylists;
}

export interface GetBrowseFeaturedPlaylistsDataPlaylists {
  href: string;
  items: Item4[];
  limit: number;
  next: string;
  offset: number;
  previous: null;
  total: number;
}

export interface Item4 {
  collaborative: boolean;
  description: string;
  external_urls: ExternalUrls;
  href: string;
  id: string;
  images: Icon[];
  name: string;
  owner: Owner;
  primary_color: PrimaryColor | null;
  public: boolean;
  snapshot_id: string;
  tracks: Followers;
  items: Followers;
  type: GetPlaylistDataType;
  uri: string;
}

export enum PrimaryColor {
  Eb1E32 = '#eb1e32',
  Ffffff = '#ffffff',
  PrimaryColorFFFFFF = '#FFFFFF',
  The509Bf5 = '#509BF5',
}

export enum GetPlaylistDataType {
  Playlist = 'playlist',
}

export interface GetBrowseNewReleasesData {
  albums: GetBrowseNewReleasesDataAlbums;
}

export interface GetBrowseNewReleasesDataAlbums {
  href: string;
  items: ItemElement[];
  limit: number;
  next: string;
  offset: number;
  previous: null;
  total: number;
}